unpinned_requirements.txt
opencv-python
pillow
matplotlib
pytest-lazy-fixture